pyproject.toml
[tool.poetry]
name = "nonebot_plugin_PCtrl"
version = "0.2.2"
description = ""
authors = ["CherryGS <uplliode@gmail.com>"]
license = "MIT"
homepage = "https://github.com/CherryGS/nonebot_plugin_PCtrl"
repository = "https://github.com/CherryGS/nonebot_plugin_PCtrl"
readme = "README.md"
documentation = "https://github.com/CherryGS/nonebot_plugin_PCtrl#readme"
packages = [
    { include = "controller", from = "." },
    { include = "__init__.py", from = "." }
]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.29"
aiosqlite = "^0.17.0"
anyutils = "^0.2.0"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.17.2"
pytest-cover = "^3.0.0"

[tool.pytest.ini_options]
testpaths = "test"
asyncio_mode="auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
